carWash -> third iteration

+++ FXCarWash/Source/FXCar/FXEmployee/FXEmployee.h
@@ -10,9 +10,9 @@
 
 @interface FXEmployee : NSObject
 @property (nonatomic, assign)	uint64_t	experience;
-@property (nonatomic, assign)	uint32_t	salary;
+@property (nonatomic, assign)	float		salary;
 
-- (id)initWithExperience:(uint64_t)experience salary:(uint32_t)salary;
+- (id)initWithExperience:(uint64_t)experience salary:(float)salary;
 
 - (id)performEmployeeSpecificJob;
