+++ FXCarWash/Source/FXCar/FXEmployee/FXEmployee.m
@@ -25,7 +25,7 @@
 	return [self initWithExperience:0 salary:0];
 }
 
-- (id)initWithExperience:(uint64_t)experience salary:(uint32_t)salary {
+- (id)initWithExperience:(uint64_t)experience salary:(float)salary {
 	self = [super init]; // init superclass
 	
 	if (self) {
